refactor(cron): extract legacy migration (#1900)
