+++ apps/server/api/src/collections/cron-jobs/utils/cron-schedule.util.ts
@@ -1,0 +1,16 @@
+import { CronJob as CronParser } from 'cron';
+
+export function computeNextRunAtOrThrow(
+  schedule: string,
+  timezone: string | undefined,
+): Date {
+  const parser = new CronParser(
+    schedule,
+    () => undefined,
+    null,
+    false,
+    timezone ?? 'UTC',
+  );
+
+  return parser.nextDate().toJSDate();
+}
